Add support for parsing IPv6 addresses
